requirements.txt
Django
gunicorn
mysqlclient
pymysql
djangorestframework
